Add list-deliveries command
To make it easier to find the delivery id to replay
